feat: add schema and trade persistence with DB integration and refactor project structure

+++ src/main/java/com/example/find_my_edge/common/dto/FilterDTO.java
@@ -6,8 +6,7 @@ import lombok.Data;
 public class FilterDTO {
     private String key;
     private String operator;
-    private int value;
-    private int from;
-    private int to;
-
+    private Double value;
+    private Double from;
+    private Double to;
 }
